Massive refactoring
Big optimizations and API changes

+++ websocket-sharp/PayloadData.cs
@@ -75,7 +75,7 @@ namespace WebSocketSharp
         static PayloadData()
         {
             Empty = new PayloadData();
-            MaxLength = long.MaxValue;
+            MaxLength = int.MaxValue;
         }
 
         #endregion
